requirements.txt
Pillow==8.4.0
streamlit==1.2.0
